Package with fish.payara prefix. Also renamed the maven artifact
