+++ cloud-providers/azure/kubernetes/payara-micro-distributed-cache/pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>fish.payara.accelerator.k8s.azure</groupId>
-    <artifactId>testapp</artifactId>
+    <artifactId>payara-azure-k8s-testapp</artifactId>
     <version>1.0</version>
     <packaging>war</packaging>
