+++ cloud-providers/azure/kubernetes/payara-micro-distributed-cache/src/main/java/fish/payara/examples/azure/k8s/UserData.java
@@ -1,4 +1,4 @@
-package be.rubus.accelerator.k8s.azure;
+package fish.payara.examples.azure.k8s;
 
 import javax.json.bind.annotation.JsonbCreator;
 import javax.json.bind.annotation.JsonbProperty;
